requirements.txt
Django==4.1.7
djangorestframework==3.14.0
model_bakery==1.12.0